+++ build.rs
@@ -125,6 +125,7 @@ fn build_capstone_cc() {
         .flag_if_supported("-Wno-sign-compare")
         .flag_if_supported("-Wno-return-type")
         .flag_if_supported("-Wno-implicit-fallthrough")
+        .flag_if_supported("-Wno-missing-field-initializers")
         .static_crt(use_static_crt)
         .compile("capstone");
 }