+++ capstone/arch/Sparc/SparcInstPrinter.c
@@ -12,7 +12,7 @@
 //===----------------------------------------------------------------------===//
 
 /* Capstone Disassembly Engine */
-/* By Nguyen Anh Quynh <aquynh@gmail.com>, 2013-2014 */
+/* By Nguyen Anh Quynh <aquynh@gmail.com>, 2013-2015 */
 
 #ifdef CAPSTONE_HAS_SPARC
 
@@ -165,7 +165,7 @@ static bool printSparcAliasInstr(MCInst *MI, SStream *O)
 
 static void printOperand(MCInst *MI, int opNum, SStream *O)
 {
-	int Imm;
+	int64_t Imm;
 	unsigned reg;
 	MCOperand *MO = MCInst_getOperand(MI, opNum);
 
@@ -200,8 +200,8 @@ static void printOperand(MCInst *MI, int opNum, SStream *O)
 		// backward, so they need to be multiplied by 4
 		switch (MI->Opcode) {
 			case SP_CALL:
-				Imm = SignExtend32(Imm, 30);
-				Imm += (uint32_t)MI->address;
+				// Imm = SignExtend32(Imm, 30);
+				Imm += MI->address;
 				break;
 
 			// Branch on integer condition with prediction (BPcc)
@@ -219,7 +219,7 @@ static void printOperand(MCInst *MI, int opNum, SStream *O)
 			case SP_BPFCCANT:
 			case SP_BPFCCNT:
 				Imm = SignExtend32(Imm, 19);
-				Imm = (uint32_t)MI->address + Imm * 4;
+				Imm = MI->address + Imm * 4;
 				break;
 
 			// Branch on integer condition (Bicc)
@@ -230,7 +230,7 @@ static void printOperand(MCInst *MI, int opNum, SStream *O)
 			case SP_FBCOND:
 			case SP_FBCONDA:
 				Imm = SignExtend32(Imm, 22);
-				Imm = (uint32_t)MI->address + Imm * 4;
+				Imm = MI->address + Imm * 4;
 				break;
 
 			// Branch on integer register with prediction (BPr)
@@ -259,11 +259,11 @@ static void printOperand(MCInst *MI, int opNum, SStream *O)
 			case SP_BPZnapn:
 			case SP_BPZnapt:
 				Imm = SignExtend32(Imm, 16);
-				Imm = (uint32_t)MI->address + Imm * 4;
+				Imm = MI->address + Imm * 4;
 				break;
 		}
 		
-		printInt32(O, Imm);
+		printInt64(O, Imm);
 
 		if (MI->csh->detail) {
 			if (MI->csh->doing_mem) {
@@ -358,8 +358,8 @@ void Sparc_printInst(MCInst *MI, SStream *O, void *Info)
 	mnem = printAliasInstr(MI, O, Info);
 	if (mnem) {
 		// fixup instruction id due to the change in alias instruction
-		strncpy(instr, mnem, strlen(mnem));
-		instr[strlen(mnem)] = '\0';
+		strncpy(instr, mnem, sizeof(instr));
+		instr[sizeof(instr) - 1] = '\0';
 		// does this contains hint with a coma?
 		p = strchr(instr, ',');
 		if (p)