+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - YYYY-MM-DD
+
+### Changed
+- Upgraded bundled capstone to release [4.0](https://github.com/aquynh/capstone/releases/tag/4.0)
+  (Git commit [db19431d to 75c05281](https://github.com/aquynh/capstone/compare/db19431d...75c05281)).
+
 ## [0.9.1] - 2018-09-20
 ### Changed
 - Upgraded bundled capstone to release [3.0.5](https://github.com/aquynh/capstone/releases/tag/3.0.5)